Bring Group Tags Administration in network Admin if BuddyPress is network activated
So far it was not possible to manage Group Tags when BuddyPress is network activated. The problem in the network Administration is that WordPress does not provide the edit-tags.php page. BuddyPress Groups, in this case are managed from the network Administration. I think having the group tags administration in the root site admin and the groups administration in the network is not great at all. So i've taken some of the code of the latest patch i have posted on the BuddyPress core ticket 4017 and applyed it to this plugin. The result is: no matter if BuddyPress is activated on the network or not, the Group Tags administration will always be a submenu of the Groups Administration.

Some other minor improvements comes with this major one such as :
- A better management of term count (hidden groups are not part of the count)
- A new way to come back to all groups when browsing a tag (a close button)