Introduce a new filter to customize the tag link
It's now possible to use the filter 'bp_group_tags_params' to add a title attribute to the tag link. This title will contain an excerpt of the description of the tag. The 2 other params are the hooks to use to output the links. By default the array of params is array( 'taglink_description' => 0, 'directory_hook' => 'bp_directory_groups_item', 'group_hook' => 'bp_before_group_header_meta' ). The filter will have to return an array with the custom values. About the taglink_description parameter, it's an integer used to inform about the length in number of words.

Fixes #1, #3