Make sure queries are built the right way for any groups filters
when filtering groups alphabetically or by date of group creation, the WHERE parts of the sql built by BuddyPress is empty. As the Tax Query WordPress function is prefixing by default the sql where clause with AND we need to remove this AND in these cases so that the query is correctly built.

Props @coldacid

Fixes #17